add PaymentScreen
